Move input handling code to a dedicated class
